kubernetes/2/app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.2.0
    # via kombu
asgiref==3.8.1
    # via django
async-timeout==4.0.3
    # via redis
billiard==4.2.0
    # via celery
celery==5.4.0
    # via -r requirements.in
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
django==5.0
    # via
    #   -r requirements.in
    #   django-redis
django-redis==5.4.0
    # via -r requirements.in
kombu==5.4.0
    # via celery
prompt-toolkit==3.0.47
    # via click-repl
psycopg2-binary==2.9.9
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via celery
python-decouple==3.8
    # via -r requirements.in
redis==5.0.8
    # via django-redis
six==1.16.0
    # via python-dateutil
sqlparse==0.5.1
    # via django
typing-extensions==4.12.2
    # via asgiref
tzdata==2024.1
    # via celery
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
